oca-port: new option '--non-interactive'
This options will only perform the analysis of the addon and stops there,
without asking the user what he wants to do.
The main purpose of this option is to be used in a script, like
performing the analysis of all the addons of a repository to get a report.

If the addon has commits to port or could be migrated to the targeted
branch, the tool will exit with a non-zero return code.