[FIX] Port orphaned commits in the right order
Commits from PR (fetched from GitHub API) had already the right order
so the porting process was working as expected.

That was not the case of orphaned commits (without PR) which were
initialized in a loop that was starting from latest commits (children)
to older ones (ancestors), so the porting process was generating a lot
of conflicts.