[FIX] ignore file paths outside of './subfolder'
An exception 'ValueError: 'path' is not in the subpath of './subfolder'"
could be raised when porting modules located in a subfolder if one of
its commit is also updating a file outside of this subfolder.

Keep things simple: ignore such file paths, we focus only on commits
located in the subfolder.

This should have no impact in OCA-like repositories as modules are all
located in the root folder (subfolder='.').