PortAddonPullRequest: reduce noise by filtering PRs to port
PRs that are updating both the analyzed module + other ones (here called
'satellite') were listed by oca-port, even if the commits related to the
analyzed module were already ported, because contributions on these
satellite modules were not (yet) ported.

In the context of the analyzed module, this brings noise and such PR
should not be proposed to be ported, but only mentioned.

There are now two kinds of detected PRs:

- the ones that are updating the current analyzed module: they are
  proposed to the user as "to be ported" as usual
- the ones that are updating only satellite modules/files:
- the ones where only satellite modules/files remain to be ported: such
  modules are now listed (in verbose mode), and the user invited to run
  oca-port on them.

As such PRs with commits updating only satellite modules are not
proposed anymore to the user, some modules will now be considered as
"Fully ported" instead of still having some "Commits to port".